[4.8] Merge Google gcc-4_8 branch
Merge head is at git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google/gcc-4_8@201445
138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc-4.8/gcc/tree-optimize.c
@@ -51,6 +51,7 @@ static unsigned int
 execute_cleanup_cfg_post_optimizing (void)
 {
   unsigned int todo = 0;
+
   if (cleanup_tree_cfg ())
     todo |= TODO_update_ssa;
   maybe_remove_unreachable_handlers ();
@@ -128,22 +129,22 @@ execute_fixup_cfg (void)
 
   if (ENTRY_BLOCK_PTR->count)
     count_scale = ((cgraph_get_node (current_function_decl)->count
-		    * REG_BR_PROB_BASE + ENTRY_BLOCK_PTR->count / 2)
+		    * (double) REG_BR_PROB_BASE + ENTRY_BLOCK_PTR->count / 2)
 		   / ENTRY_BLOCK_PTR->count);
   else
     count_scale = REG_BR_PROB_BASE;
 
   ENTRY_BLOCK_PTR->count = cgraph_get_node (current_function_decl)->count;
-  EXIT_BLOCK_PTR->count = (EXIT_BLOCK_PTR->count * count_scale
+  EXIT_BLOCK_PTR->count = (EXIT_BLOCK_PTR->count * (double) count_scale
   			   + REG_BR_PROB_BASE / 2) / REG_BR_PROB_BASE;
 
   FOR_EACH_EDGE (e, ei, ENTRY_BLOCK_PTR->succs)
-    e->count = (e->count * count_scale
+    e->count = (e->count * (double) count_scale
        + REG_BR_PROB_BASE / 2) / REG_BR_PROB_BASE;
 
   FOR_EACH_BB (bb)
     {
-      bb->count = (bb->count * count_scale
+      bb->count = (bb->count * (double) count_scale
 		   + REG_BR_PROB_BASE / 2) / REG_BR_PROB_BASE;
       for (gsi = gsi_start_bb (bb); !gsi_end_p (gsi); gsi_next (&gsi))
 	{
@@ -177,7 +178,7 @@ execute_fixup_cfg (void)
 	}
 
       FOR_EACH_EDGE (e, ei, bb->succs)
-        e->count = (e->count * count_scale
+          e->count = (e->count * (double) count_scale
 		    + REG_BR_PROB_BASE / 2) / REG_BR_PROB_BASE;
 
       /* If we have a basic block with no successors that does not
